partial changes for the lease escalations to be fixed in case of subsequent modification

+++ app/Http/Controllers/Drafts/IndexController.php
@@ -9,6 +9,7 @@
 namespace App\Http\Controllers\Drafts;
 
 use App\Http\Controllers\Controller;
+use App\LeaseHistory;
 use Illuminate\Http\Request;
 use App\Lease;
 use App\LeaseAssets;
@@ -54,12 +55,21 @@ class IndexController extends Controller
     {
         try {
             if ($request->ajax()) {
-                return datatables()->eloquent(Lease::query()->whereIn('business_account_id', getDependentUserIds())->where('status', '=', '0')->with('assets'))->toJson();
+                return datatables()->eloquent(
+                    Lease::query()->whereIn('business_account_id', getDependentUserIds())
+                        ->where('status', '=', '0')->with('assets')
+                )
+                    ->addColumn('can_be_deleted', function ($data) {
+                        //a lease from drafts cannot be deleted if it has been submitted at lease once.
+                        $history_exist = LeaseHistory::query()->where('lease_id', '=', $data->id)->count();
+                        return $history_exist < 1;
+                    })
+                    ->toJson();
             } else {
                 return redirect()->back();
             }
         } catch (\Exception $e) {
-            dd($e);
+            abort(404);
         }
     }
 
@@ -75,8 +85,13 @@ class IndexController extends Controller
             if ($request->ajax()) {
                 $lease = Lease::query()->whereIn('business_account_id', getDependentUserIds())->where('id', $id)->first();
                 if ($lease) {
-                    $lease->delete();
-                    return response()->json(['status' => true], 200);
+                    $history_exist = LeaseHistory::query()->where('lease_id', '=', $lease->id)->count();
+                    if($history_exist < 1){
+                        $lease->delete();
+                        return response()->json(['status' => true], 200);
+                    } else {
+                        return response()->json(['status' => false, "message" => "You have already submitted this lease and hence cannot be deleted."], 200);
+                    }
                 } else {
                     return response()->json(['status' => false, "message" => "Invalid request!"], 200);
                 }
